Making features vectors
To be efficient and to ensure that all feature vectors have the same format, we should:
First, compute the list of features.
Then, tally the heuristics in a dict as we make a pass through the tree.

Unfortunately, I am having problems with this because the parse and fuzz methods of some NonTerminals in particular are incorrect in that they are producing parent types rather than objects of the same class.
To fix this, I'm adding an interface to DataModel which allows us to propagate an object, specifying flexible details on how it should differ from its parent (creation-wise, not type-wise) object.